Moved Java files into two separate folders. One for the Database related classes and once for the GUI related activity classes.
