+++ app/src/main/java/com/example/pillboxesapp/Activities/EmailChecker.java
@@ -1,14 +1,14 @@
-package com.example.pillboxesapp;
+package com.example.pillboxesapp.Activities;
 
 import java.util.regex.Pattern;
 
-public class emailChecker {
+public class EmailChecker {
 
     protected static final String EMAIL_FORMAT_ERROR = "Email not formatted correctly";
     protected static final String EMAIL_EMPTY_ERROR = "Email required";
     private String emailAddress;
 
-    public emailChecker(String email) {
+    public EmailChecker(String email) {
         emailAddress = email;
     }
