+++ app/src/main/java/com/example/pillboxesapp/Activities/ForgotPasswordActivity.java
@@ -1,4 +1,4 @@
-package com.example.pillboxesapp;
+package com.example.pillboxesapp.Activities;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -13,6 +13,7 @@ import android.widget.EditText;
 import android.widget.ProgressBar;
 import android.widget.Toast;
 
+import com.example.pillboxesapp.R;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
@@ -90,12 +91,12 @@ public class ForgotPasswordActivity extends AppCompatActivity {
 
     private boolean validFields(String email) {
         if (TextUtils.isEmpty(email)) {
-            emailBox.setError(emailChecker.EMAIL_EMPTY_ERROR);
+            emailBox.setError(EmailChecker.EMAIL_EMPTY_ERROR);
             return false;
         }
-        emailChecker emailCheck = new emailChecker(email);
+        EmailChecker emailCheck = new EmailChecker(email);
         if (!emailCheck.isEmailAddressValid()) {
-            emailBox.setError(emailChecker.EMAIL_FORMAT_ERROR);
+            emailBox.setError(EmailChecker.EMAIL_FORMAT_ERROR);
             return false;
         }
         return true;
