+++ app/src/main/java/com/example/pillboxesapp/Database/CallbackResult.java
@@ -1,4 +1,4 @@
-package com.example.pillboxesapp;
+package com.example.pillboxesapp.Database;
 
 import java.util.ArrayList;
