+++ app/src/main/java/com/example/pillboxesapp/Database/DatabaseService.java
@@ -1,4 +1,4 @@
-package com.example.pillboxesapp;
+package com.example.pillboxesapp.Database;
 
 import androidx.annotation.NonNull;
