+++ app/src/main/java/com/example/pillboxesapp/Database/UserEntity.java
@@ -1,4 +1,4 @@
-package com.example.pillboxesapp;
+package com.example.pillboxesapp.Database;
 
 public class UserEntity {
     private String userID;
