+++ app/src/main/res/layout/activity_forgot_password.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/colorPrimaryDark"
     android:fillViewport="true"
-    tools:context=".ForgotPasswordActivity">
+    tools:context=".Activities.ForgotPasswordActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
